kv488: set up skeleton
Summary:
This commit adds files for each movement and instrument, so it should be
the case that all that’s left to do is to fill in the notes!

Test Plan:
Check that compilation is successful:

```
$ cd ftp/MozartWA/KV488/
$ ! lilypond MozartWA-KV488-*.ly 2>&1 | grep -e warn -e err
$ ls -1 MozartWA-KV488-*.pdf | wc -l
10
$ ls -1 *.midi | wc -l
3
```

Verify in the conductor score that all parts are present, and verify
that the clarinet and horn parts are transposed to A in both the
individual part scores and the main score.