kv488: add short instrument names to all systems
Summary:
In some cases, it may be ambiguous whether the clarinet or the French
horn is introduced into a new system. The ambiguity only exists for this
particular pair (all other pairs are either far enough apart in the
staff order or have differentiating clef or transposition), but even
when a new instrument is unambiguous, it may not be obvious. This commit
adds short instrument names to each system.

The solution is a bit heavy-handed, adding visual clutter to all systems
even when in the vast majority of cases they do not add value. For
reference, the original score for the Allegro labels the instruments of
exactly two non-initial systems: those beginning at measures 265 and
271, both on page 19. I’m open to a better solution, and I’m also open
to reverting this entirely.