kv488: separate per-movement `global.ily` files
Summary:
This commit restructures the `music.ily` files such that they only
contain `\include` directives. This enables their dependents to instead
depend on only a more precise subset of the `\include`d files, reducing
the scope of the dependency.

Test Plan:
Running `make test` passes, and the output is unchanged.