kv488: match layout staff size to global size
Summary:
Because `layout-set-staff-size` does not adjust staff line spacing, the
output looks quite bad when the two values differ. If the global staff
size is set to be smaller than the layout staff size (as was the case
prior to this commit), then notes on adjacent lines/spaces will overlap
each other. This is particularly visible and ugly with whole note chords
like `<c e g>1`. Quarter-note chords look bad, too.

In this commit, we change the layout staff size from 17 to 14 to match
the global staff size. Adjusting the global staff size would require
recomputing the dimensions for the cover page and table of contents,
so we opt for the path of least resistance for now.