Changes to update the LilyPond version from 2.20.33 to 2.24.
By comparing the original PDF to the new output, I was able to make minimal changes to the input file while still keeping the output true to its original form.  I added a \paper block and some staff-staff-spacing because the output was too crowded but beyond that, very little was done other than the convertion.